requirements.txt
edge-tts==7.2.7
face-recognition==1.3.0
openai==2.26.0
opencv-python==4.13.0.92
pyTelegramBotAPI==4.32.0
python-dotenv==1.2.2
requests==2.32.5
SpeechRecognition==3.14.5
apscheduler==3.11.2
Pillow==12.1.1
